Moved plugin subclasses to separate headers
