+++ licq/include/licq/pluginmanager.h
@@ -1,6 +1,6 @@
 /*
  * This file is part of Licq, an instant messaging client for UNIX.
- * Copyright (C) 2010 Licq developers
+ * Copyright (C) 2010-2011 Licq developers
  *
  * Licq is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,7 +24,8 @@
 #include <list>
 #include <string>
 
-#include "plugin.h"
+#include "generalplugin.h"
+#include "protocolplugin.h"
 
 
 namespace Licq
