+++ licq/src/plugins/generalplugin.h
@@ -1,6 +1,6 @@
 /*
  * This file is part of Licq, an instant messaging client for UNIX.
- * Copyright (C) 2010 Licq developers
+ * Copyright (C) 2010-2011 Licq developers
  *
  * Licq is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -21,6 +21,7 @@
 #define LICQDAEMON_GENERALPLUGIN_H
 
 #include "plugin.h"
+#include <licq/generalplugin.h>
 
 #include <list>
 #include <queue>
