+++ qt4-gui/src/dialogs/plugindlg.h
@@ -1,6 +1,6 @@
 /*
  * This file is part of Licq, an instant messaging client for UNIX.
- * Copyright (C) 1999-2010 Licq developers
+ * Copyright (C) 1999-2011 Licq developers
  *
  * Licq is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -20,7 +20,8 @@
 #ifndef PLUGINDLG_H
 #define PLUGINDLG_H
 
-#include <licq/plugin.h>
+#include <licq/generalplugin.h>
+#include <licq/protocolplugin.h>
 
 #include <QMap>
 #include <QDialog>
